+++ runelite-client/src/main/java/net/runelite/client/plugins/coxhelper/CoxInfoBox.java
@@ -36,7 +36,6 @@ import net.runelite.api.Client;
 import net.runelite.api.NpcID;
 import net.runelite.api.SpriteID;
 import net.runelite.client.game.SpriteManager;
-import net.runelite.client.plugins.coxhelper.utils.PrayAgainst;
 import net.runelite.client.ui.overlay.Overlay;
 import net.runelite.client.ui.overlay.OverlayPosition;
 import net.runelite.client.ui.overlay.components.ComponentConstants;
