+++ runelite-client/src/main/java/net/runelite/client/plugins/coxhelper/CoxPlugin.java
@@ -65,7 +65,6 @@ import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.plugins.PluginType;
-import net.runelite.client.plugins.coxhelper.utils.PrayAgainst;
 import net.runelite.client.ui.overlay.OverlayManager;
 import net.runelite.client.util.Text;
