+++ runelite-client/src/main/java/net/runelite/client/plugins/coxhelper/PrayAgainst.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.coxhelper.utils;
+package net.runelite.client.plugins.coxhelper;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
